note that the recon/reflow/reTurn on this branch is completely out of sync with main; it does not include the sipXmedia dependencies, nor does it include any other media stack, however it does provide interfaces that you can implement so as to plug in your own media stack

git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/cp-recon-contrib-20100125@9213 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ contrib/ares/ares__read_line.c
@@ -45,7 +45,7 @@ int ares__read_line(FILE *fp, char **buf, int *bufsize)
     {
       if (!fgets(*buf + offset, *bufsize - offset, fp))
 	return (offset != 0) ? 0 : (ferror(fp)) ? ARES_EFILE : ARES_EOF;
-      len = offset + strlen(*buf + offset);
+      len = offset + (int)strlen(*buf + offset);
       if ((*buf)[len - 1] == '\n')
 	{
 	  (*buf)[len - 1] = 0;
